added new create event component
