+++ volunteer-app/src/app/app.component.html
@@ -1,1 +1,2 @@
 <app-login></app-login>
+<app-create-event></app-create-event>
