+++ volunteer-app/src/app/app.module.ts
@@ -7,6 +7,7 @@ import { AppComponent } from './app.component';
 import { ToolbarComponent } from './toolbar/toolbar.component';
 import { LoginComponent } from './login/login.component';
 import { SignupComponent } from './signup/signup.component';
+import { CreateEventComponent } from './create-event/create-event.component';
 
 var firebaseConfig = {
   apiKey: "AIzaSyAxq4Bb-wcQ3Hmg-8lqMqaXciTpd699Jw0",
@@ -23,7 +24,8 @@ var firebaseConfig = {
     AppComponent,
     ToolbarComponent,
     LoginComponent,
-    SignupComponent
+    SignupComponent,
+    CreateEventComponent
   ],
   imports: [
     AngularFireModule.initializeApp(firebaseConfig),
