--- volunteer-app/src/app/calandar/calandar.component.html
@@ -1,1 +1,0 @@
-<p>calandar works!</p>
