--- volunteer-app/src/app/calandar/calandar.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-calandar',
-  templateUrl: './calandar.component.html',
-  styleUrls: ['./calandar.component.css']
-})
-export class CalandarComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
