+++ volunteer-app/src/app/create-event/create-event.component.html
@@ -1,0 +1,14 @@
+<p>create-event works!</p>
+
+<h3>Create event</h3>
+<h4> 
+  <input id="name" placeholder="event name"/> 
+  <input id="category" placeholder="category"/> 
+  <input id="date" placeholder="YYYY-MM-DD"/> 
+  <input id="startTime" placeholder="Start Time : HH:MM:SS"/> 
+  <input id="endTime" placeholder="End Time : HH:MM:SS"/> 
+  <input id="organizer" placeholder="organizer name"/> 
+  
+  <button 
+  (click)="createEvent()"> Create event</button></h4>
+
