+++ volunteer-app/src/app/create-event/create-event.component.spec.ts
@@ -1,20 +1,20 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { CalandarComponent } from './calandar.component';
+import { CreateEventComponent } from './create-event.component';
 
-describe('CalandarComponent', () => {
-  let component: CalandarComponent;
-  let fixture: ComponentFixture<CalandarComponent>;
+describe('CreateEventComponent', () => {
+  let component: CreateEventComponent;
+  let fixture: ComponentFixture<CreateEventComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ CalandarComponent ]
+      declarations: [ CreateEventComponent ]
     })
     .compileComponents();
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(CalandarComponent);
+    fixture = TestBed.createComponent(CreateEventComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
