+++ volunteer-app/src/app/login/login.component.html
@@ -4,16 +4,3 @@
 <h4>{{ user.first_name }} {{ user.last_name }} 
   <input id="a"/> <button 
   (click)="updateJohn()"> Update Volunteer</button></h4>
-
-<h3>Create event</h3>
-<h4> 
-  <input id="name" placeholder="event name"/> 
-  <input id="category" placeholder="category"/> 
-  <input id="date" placeholder="YYYY-MM-DD"/> 
-  <input id="startTime" placeholder="Start Time : HH:MM:SS"/> 
-  <input id="endTime" placeholder="End Time : HH:MM:SS"/> 
-  <input id="organizer" placeholder="organizer name"/> 
-  
-  <button 
-  (click)="createEvent()"> Create event</button></h4>
-
